pyproject.toml
[tool.poetry]
name = "brain-mri-classification"
version = "0.1.0"
description = ""
authors = ["Joanna Khek <joannakhek@gmail.com>"]
readme = "README.md"
packages = [{include = "cnn_model", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.10"
pandas = ">=1.3.5"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
kaggle = "^1.5.16"
pyyaml = "^6.0.1"
numpy = "^1.21.4"
torch = {version = "^2.0.1+cu118", source = "pytorch-gpu-src"}
torchvision = {version = "^0.15.2+cu118", source = "pytorch-gpu-src"}
torchaudio = {version = "^2.0.2+cu118", source = "pytorch-gpu-src"}
torchinfo = "^1.8.0"
scikit-learn = "^1.3.2"
torchmetrics = "^1.2.1"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipykernel = "^6.25.2"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
